Refactor hammer_vlsi into multiple files
About time, since hammer_vlsi.py was getting really large (2k+ LoC!).

In addition, mypy has trouble with import cycles:
https://github.com/python/mypy/issues/481
